Move component to isomorphic and move function buildComponentFromVNode for createComponent.js
